--- dl4j-examples/src/main/resources/import/tensorflow/input_a.shape
@@ -1,2 +1,0 @@
-1
-784
