--- dl4j-examples/src/main/resources/import/tensorflow/input_b.shape
@@ -1,2 +1,0 @@
-1
-784
