--- dl4j-examples/src/main/resources/import/tensorflow/prediction_b.csv
@@ -1,10 +1,0 @@
-0.00000000
-0.00000000
-1.00000000
-0.00000000
-0.00000000
-0.00000000
-0.00000000
-0.00000000
-0.00000000
-0.00000000
