--- dl4j-examples/src/main/resources/import/tensorflow/prediction_b.shape
@@ -1,2 +1,0 @@
-1
-10
