+++ dl4j-examples/src/main/resources/modelimport/tensorflow/input_a.shape
@@ -1,0 +1,2 @@
+1
+784
