+++ dl4j-examples/src/main/resources/modelimport/tensorflow/prediction_a.csv
@@ -1,0 +1,10 @@
+0.00000000
+0.00000000
+0.00000000
+0.00000000
+0.00000000
+0.00000000
+0.00000000
+1.00000000
+0.00000000
+0.00000000
