+++ dl4j-examples/src/main/resources/modelimport/tensorflow/prediction_b.shape
@@ -1,0 +1,2 @@
+1
+10
